Compiled .jar, changed action of cancel button on password entry.
Version has been updatated to 0.3.0. This marks the transition out of the second stable alpha build, and into the third stable alpha build. Until alpha build 0.4, this will be your most stable build. The server numbering has been updated to support this change.

The cancel button on password entry now quits the program instead of prompting the user to enter a password again.